nvme-types: move mi definitions into separate header
The nvme-types.h header has become too large and difficult to maintain.
