+++ libnvme/libnvme/nvme.i
@@ -1376,4 +1376,5 @@ struct libnvme_ns *libnvme_ctrl_next_ns(struct libnvme_ctrl *c, struct libnvme_n
 
 %include "../src/nvme/nvme-types.h"
 %include "../src/nvme/nvme-types-fabrics.h"
+%include "../src/nvme/nvme-types-mi.h"
 %include "../src/nvme/nvme-types-zns.h"
